Accessing name node URL with portnumber 500070
Here the url for namenode from the previous screenshot can be determined as "nn01.itversity.com", so try the port number 50070 with the URL as follows:

nn01.itversity.com:50070

Once this URL is activated with port 50070 you can see an overview of the namenode. 

This screenshot displays the configured Capacity. Further screenshots will show information when scrolled down.