Finding namenodes IP address when it is not given
This will be useful while taking certifications on Hadoop/Spark. You will have only the link to the domain of the cluster. They will not have specified the namenode UI link or the namenodes IP address. 

Using the following steps you can get the IP address of the namenode from console

1. Navigate to /etc/hadoop/conf/ (cd /etc/hadoop/conf/)
2. ls -ltr 
3. ls -ltr core-site.xml hdfs-site.xml (specifically looking for these files)
4. vim core-site.xml
5. Search for the property "fs.defaultFS" . The value is the URL for the namenode UI. This is the place where master daemon process / name node process runs
6. using this namenode URL add port number 50070 to access the webservice on namenode UI
7. if you look at hdfs-site.xml (vim hdfs-site.xml) -> important properties are dfs.blocksize and dfs.replication (to search in the file use / followed by the search string, for eg /replication